+++ v2/rest/lists/PinnedLists.ts
@@ -1,4 +1,4 @@
-import type { Snowflake } from '../..';
+import type { Snowflake } from '../../payloads';
 import type { MultipleListsLookupQuery, MultipleListsLookupResponse } from './ListsLookup';
 
 /**
