+++ v2/rest/misc.ts
@@ -1,18 +1,18 @@
 import type { APIList, APIMedia, APIPlace, APIPoll, APISpace, APITweet, APIUser } from '../payloads';
 
-export type ListFieldsParameter = keyof APIList;
+export type APIListFieldsParameter = keyof APIList;
 
-export type MediaFieldsParameter = keyof APIMedia;
+export type APIMediaFieldsParameter = keyof APIMedia;
 
-export type PlaceFieldsParameter = keyof APIPlace;
+export type APIPlaceFieldsParameter = keyof APIPlace;
 
-export type PollFieldsParameter = keyof APIPoll;
+export type APIPollFieldsParameter = keyof APIPoll;
 
-export type SpaceFieldsParameter = keyof APISpace;
+export type APISpaceFieldsParameter = keyof APISpace;
 
-export type TweetFieldsParameter = keyof APITweet;
+export type APITweetFieldsParameter = keyof APITweet;
 
-export type UserFieldsParameter = keyof APIUser;
+export type APIUserFieldsParameter = keyof APIUser;
 
 // eslint-disable-next-line no-unused-vars
 export type Without<FirstType, SecondType> = { [KeyType in Exclude<keyof FirstType, keyof SecondType>]?: never };
