+++ v2/rest/tweets/Likes.ts
@@ -1,6 +1,6 @@
-import type { MultipleTweetsLookupQuery, MultipleTweetsLookupResponse } from '.';
-import type { Snowflake } from '../..';
-import type { MultipleUsersLookupQuery, MultipleUsersLookupResponse } from '../users/UsersLookup';
+import type { Snowflake } from '../../payloads';
+import type { MultipleUsersLookupQuery, MultipleUsersLookupResponse } from '../users';
+import type { MultipleTweetsLookupQuery, MultipleTweetsLookupResponse } from './TweetsLookup';
 
 /**
  * The query for fetching users who liked a tweet
