+++ v2/rest/tweets/SearchTweets.ts
@@ -1,5 +1,5 @@
-import type { MultipleTweetsLookupQuery, MultipleTweetsLookupResponse } from '.';
-import type { Snowflake } from '../..';
+import type { Snowflake } from '../../payloads';
+import type { MultipleTweetsLookupQuery, MultipleTweetsLookupResponse } from './TweetsLookup';
 
 /**
  * The query for fetching recent tweets by search query
